c23

+++ backend/includes/list.h
@@ -12,14 +12,14 @@ void list_grow(list_t *l);
 void list_append(list_t *l, void *element);
 
 /* Puts `element` at the specified `index`. Returns -1 if theres any problem */
-int list_insert(list_t *l, void *element, int index);
+int list_insert(list_t *l, void *element, size_t index);
 
-/* Frees element and sets it to NULL */
+/* Frees element and sets it to nullptr */
 void list_default_callback(void *element);
 
 /* Removes element at the `index` value. Returns -1 if theres any problem, if
  * necesary you can use the callback function to free the element */
-int list_remove(list_t *l, int index, void (*callback)(void *));
+int list_remove(list_t *l, size_t index, void (*callback)(void *));
 
 /* Removes last element. Returns -1 if theres any problem, if necesary you can
  * use the callback function to free the element */
@@ -27,7 +27,7 @@ int list_pop(list_t *l, void (*callback)(void *));
 
 /* you should free all element of the array if necesary you can use the callback
  * function in order to do so... the array itself is freed by the function(also
- * it set them NULL)
+ * it set them nullptr)
  */
 void list_free(list_t *l, void (*callback)(void *));
